Cleanup TypeSymbolParser by moving extension method candidates into own class.
