+++ CodeCaster.SerializeThis.Serialization.Roslyn/CodeCaster.SerializeThis.Serialization.Roslyn.csproj
@@ -61,6 +61,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TypeSymbolExtensions.cs" />
     <Compile Include="TypeSymbolParser.cs" />
   </ItemGroup>
   <ItemGroup>
